refactor Burn to case class
