+++ src/test/scala/it/unibo/skalamon/model/status/NonVolatileStatusTest.scala
@@ -1,6 +1,7 @@
 package it.unibo.skalamon.model.status
 
 import it.unibo.skalamon.model.pokemon.Stat.*
+import it.unibo.skalamon.model.status.nonVolatileStatus.*
 import it.unibo.skalamon.model.pokemon.{BattlePokemon, PokemonTestUtils}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
@@ -10,7 +11,7 @@ class NonVolatileStatusTest extends AnyFlatSpec with should.Matchers:
   private val initialTurn: Int = 0
 
   private val assignedBurn: AssignedStatus[NonVolatileStatus] =
-    AssignedStatus(Burn, initialTurn)
+    AssignedStatus(Burn(), initialTurn)
   private val assignedParalyze: AssignedStatus[NonVolatileStatus] =
     AssignedStatus(Paralyze, initialTurn)
 
@@ -36,7 +37,7 @@ class NonVolatileStatusTest extends AnyFlatSpec with should.Matchers:
 
     assignedBurn.status.executeEffect(
       pokemon
-    ).currentHP shouldEqual (pokemon.currentHP - pokemon.base.hp / Burn.BurnDamageReduction)
+    ).currentHP shouldEqual (pokemon.currentHP - pokemon.base.hp / Burn.DamageReduction)
 
   "Burn" should "halve attack stat" in:
     val pokemon: BattlePokemon = PokemonTestUtils.simplePokemon4.copy(
@@ -46,8 +47,7 @@ class NonVolatileStatusTest extends AnyFlatSpec with should.Matchers:
     assignedBurn.status.executeEffect(
       pokemon
     ).base.baseStats.base(Attack) shouldEqual
-      (pokemon.base.baseStats.base(Attack) / Burn.BurnAttackReduction)
-
+      (pokemon.base.baseStats.base(Attack) / Burn.AttackReduction)
   "Paralyze" should "halve speed stat" in:
     val pokemon: BattlePokemon = PokemonTestUtils.simplePokemon4.copy(
       nonVolatileStatus = Option(assignedParalyze)
